+++ tests/e2e/core-tests/specs/index.js
@@ -15,10 +15,12 @@ const runCartPageTest = require( './shopper/front-end-cart.test' );
 const runCheckoutApplyCouponsTest = require( './shopper/front-end-checkout-coupons.test');
 const runCheckoutPageTest = require( './shopper/front-end-checkout.test' );
 const runMyAccountPageTest = require( './shopper/front-end-my-account.test' );
+const runMyAccountPayOrderTest = require( './shopper/front-end-my-account-pay-order.test' );
 const runSingleProductPageTest = require( './shopper/front-end-single-product.test' );
 const runVariableProductUpdateTest = require( './shopper/front-end-variable-product-updates.test' );
 
 // Merchant tests
+const runAddNewShippingZoneTest = require ( './merchant/wp-admin-settings-shipping-zones.test' );
 const runCreateCouponTest = require( './merchant/wp-admin-coupon-new.test' );
 const runCreateOrderTest = require( './merchant/wp-admin-order-new.test' );
 const runEditOrderTest = require( './merchant/wp-admin-order-edit.test' );
@@ -55,12 +57,14 @@ const runShopperTests = () => {
 	runCheckoutApplyCouponsTest();
 	runCheckoutPageTest();
 	runMyAccountPageTest();
+	runMyAccountPayOrderTest();
 	runSingleProductPageTest();
 	runVariableProductUpdateTest();
 };
 
 const runMerchantTests = () => {
 	runOrderSearchingTest();
+	runAddNewShippingZoneTest();
 	runCreateCouponTest();
 	runCreateOrderTest();
 	runEditOrderTest();
@@ -99,6 +103,7 @@ module.exports = {
 	runCheckoutApplyCouponsTest,
 	runCheckoutPageTest,
 	runMyAccountPageTest,
+	runMyAccountPayOrderTest,
 	runSingleProductPageTest,
 	runVariableProductUpdateTest,
 	runShopperTests,
@@ -119,6 +124,7 @@ module.exports = {
 	runMerchantOrderEmailsTest,
 	runMerchantTests,
 	runOrderSearchingTest,
+	runAddNewShippingZoneTest,
 	runProductBrowseSearchSortTest,
 	runApiTests,
 };