+++ tests/e2e/utils/src/components.js
@@ -461,6 +461,48 @@ const createCoupon = async ( couponAmount = '5', discountType = 'Fixed cart disc
 };
 
 /**
+ * Adds a shipping zone along with a shipping method.
+ *
+ * @param zoneName Shipping zone name.
+ * @param zoneLocation Shiping zone location. Defaults to United States (US).
+ * @param zipCode Shipping zone zip code. Defaults to empty one space.
+ * @param zoneMethod Shipping method type. Defaults to flat_rate (use also: free_shipping or local_pickup)
+ */
+const addShippingZoneAndMethod = async ( zoneName, zoneLocation = 'United States (US)', zipCode = ' ', zoneMethod = 'flat_rate' ) => {
+	await merchant.openNewShipping();
+	
+	// Fill shipping zone name
+	await page.waitForSelector('input#zone_name');
+	await expect(page).toFill('input#zone_name', zoneName);
+
+	// Select shipping zone location
+	// (.toSelect is not best option here because a lot of &nbsp are present in country/state names)
+	await expect(page).toFill('#zone_locations', zoneLocation);
+	await page.waitFor(1000); // avoiding flakiness
+	await page.keyboard.press('Tab');
+	await page.waitFor(1000); // avoiding flakiness
+	await page.keyboard.press('Enter');
+
+	// Fill shipping zone postcode if needed otherwise just put empty space
+	await page.waitForSelector('a.wc-shipping-zone-postcodes-toggle');
+	await expect(page).toClick('a.wc-shipping-zone-postcodes-toggle');
+	await expect(page).toFill('#zone_postcodes', zipCode);
+	await expect(page).toMatchElement('#zone_postcodes', zipCode);
+	await expect(page).toClick('button#submit');
+
+	// Add shipping zone method
+	await page.waitFor(2000); // avoiding flakiness
+	await expect(page).toClick('button.wc-shipping-zone-add-method', {text:'Add shipping method'});
+	await page.waitFor(2000); // avoiding flakiness
+	await page.waitForSelector('.wc-shipping-zone-method-description');
+	await expect(page).toSelect('select[name="add_method_id"]', zoneMethod);
+	await page.waitFor(1000); // avoiding flakiness
+	await expect(page).toClick('button#btn-ok');
+	await page.waitForSelector('#zone_locations');
+	await page.waitFor(1000); // avoiding flakiness
+};
+
+/**
  * Click the Update button on the order details page.
  *
  * @param noticeText The text that appears in the notice after updating the order.
@@ -471,7 +513,7 @@ const clickUpdateOrder = async ( noticeText, waitForSave = false ) => {
 		await page.waitFor( 2000 );
 	}
 
-	// PUpdate order
+	// Update order
 	await expect( page ).toClick( 'button.save_order' );
 	await page.waitForSelector( '.updated.notice' );
 
@@ -505,6 +547,7 @@ export {
 	verifyAndPublish,
 	addProductToOrder,
 	createCoupon,
+	addShippingZoneAndMethod,
 	createSimpleProductWithCategory,
 	clickUpdateOrder,
 	deleteAllEmailLogs,