+++ tests/e2e/utils/src/flows/merchant.js
@@ -18,7 +18,8 @@ const {
 	WP_ADMIN_PERMALINK_SETTINGS,
 	WP_ADMIN_PLUGINS,
 	WP_ADMIN_SETUP_WIZARD,
-	WP_ADMIN_WC_SETTINGS
+	WP_ADMIN_WC_SETTINGS,
+	WP_ADMIN_NEW_SHIPPING_ZONE
 } = require( './constants' );
 
 const baseUrl = config.get( 'url' );
@@ -170,11 +171,17 @@ const merchant = {
 		}
 	},
 
+	openNewShipping: async () => {
+		await page.goto( WP_ADMIN_NEW_SHIPPING_ZONE, {
+			waitUntil: 'networkidle0',
+		} );
+	},
+
 	openEmailLog: async () => {
 		await page.goto( `${baseUrl}wp-admin/tools.php?page=wpml_plugin_log`, {
 			waitUntil: 'networkidle0',
 		} );
-	}
+	},
 };
 
 module.exports = merchant;