+++ tests/e2e/core-tests/CHANGELOG.md
@@ -19,11 +19,13 @@
 - Merchant Product Edit tests
 - Merchant Product Search tests
 - Shopper Single Product tests
+- Shopper My Account Pay Order
 - Shopper Checkout Apply Coupon
 - Shopper Shop Browse Search Sort
 - Merchant Orders Customer Checkout Page
 - Shopper Cart Apply Coupon
 - Merchant Order Searching
+- Merchant Settings Shipping Zones
 - Shopper Variable product info updates on different variations
 - Merchant order emails flow